macros: rename module cql to value
